+++ relaic.cpp
@@ -3,25 +3,25 @@
 #include "nrutil.h"
 #include "funct.h"
 
-void ReLAIC(int ReAtNu,float **ReCoor,float LaVdWR,float *ReMaxC,float *ReMinC,
+void ReLAIC(int ReAtNu,double **ReCoor,double LaVdWR,double *ReMaxC,double *ReMinC,
             int *CubNum,int ****CubFAI,int ****CubLAI,int **CubLiA)
-/* This function constructs the list of receptor atoms which are in the cubes 
+/* This function constructs the list of receptor atoms which are in the cubes
    of a grid for the checking of bumps in CoNuBu :
-   CubNum  the number of cubes along each direction (1->x,2->y,3->z) 
+   CubNum  the number of cubes along each direction (1->x,2->y,3->z)
    CubFAI  index of the first atoms in each cube
    CubLAI  index of the last atoms in each cube
    CubLiA  list of atoms in each cube
-   LiAInd  index of the atom list 
+   LiAInd  index of the atom list
    CubAtX  cube number along x in which each receptor atom is
    CubAtY  cube number along y in which each receptor atom is
-   CubAtZ  cube number along z in which each receptor atom is 
+   CubAtZ  cube number along z in which each receptor atom is
    NuAtoC  number of atoms in the current cube */
 {
   int i,j,k,l,LiAInd,*CubAtX,*CubAtY,*CubAtZ,*CubLiA_L,NuAtoC,
       ***CubFAI_L,***CubLAI_L;
 
 /* Find the number of cubes along x, y and z axis */
-  for (i=1;i<=3;i++) 
+  for (i=1;i<=3;i++)
     CubNum[i]=ffloor((ReMaxC[i]-ReMinC[i])/(2*LaVdWR))+1;
 
   CubFAI_L=i3tensor(1,CubNum[1],1,CubNum[2],1,CubNum[3]);
@@ -41,7 +41,7 @@ void ReLAIC(int ReAtNu,float **ReCoor,float LaVdWR,float *ReMaxC,float *ReMinC,
     CubAtZ[i]=ffloor((ReCoor[i][3]-ReMinC[3])/(2*LaVdWR))+1;
   }
 
-/* Find the list of atoms which are in each cube and the indexes for the first 
+/* Find the list of atoms which are in each cube and the indexes for the first
    and last atoms */
   LiAInd=1;
 
@@ -54,7 +54,7 @@ void ReLAIC(int ReAtNu,float **ReCoor,float LaVdWR,float *ReMaxC,float *ReMinC,
 
         for (l=1;l<=ReAtNu;l++) {
 
-          if ((CubAtX[l]==i)&&(CubAtY[l]==j)&&(CubAtZ[l]==k)) {            
+          if ((CubAtX[l]==i)&&(CubAtY[l]==j)&&(CubAtZ[l]==k)) {
             NuAtoC=NuAtoC+1;
             CubLiA_L[LiAInd]=l;
             LiAInd=LiAInd+1;
@@ -66,7 +66,7 @@ void ReLAIC(int ReAtNu,float **ReCoor,float LaVdWR,float *ReMaxC,float *ReMinC,
           CubFAI_L[i][j][k]=0;
           CubLAI_L[i][j][k]=0;
         }
-        else 
+        else
           CubLAI_L[i][j][k]=LiAInd-1;
 
       }
