+++ relaic_en.cpp
@@ -3,20 +3,20 @@
 #include "nrutil.h"
 #include "funct.h"
 
-void ReLAIC_en(int ReReNu,float **ReCoor,float GrSiCu_en,float *ReMinC,
+void ReLAIC_en(int ReReNu,double **ReCoor,double GrSiCu_en,double *ReMinC,
                int *CubNum_en,int ***CubFAI_en,int ***CubLAI_en,int *CubLiA_en,
                int *AtReprRes,FILE *FPaOut)
-/* This function constructs the list of receptor residues which are in the 
-   cubes of a grid for the energy evaluation. The position of each residue is 
+/* This function constructs the list of receptor residues which are in the
+   cubes of a grid for the energy evaluation. The position of each residue is
    considered to be the position of its residue-representative atom :
-   CubNum_en  the number of cubes along each direction (1->x,2->y,3->z) 
+   CubNum_en  the number of cubes along each direction (1->x,2->y,3->z)
    CubFAI_en  index of the first residues in each cube
    CubLAI_en  index of the last residues in each cube
    CubLiA_en  list of residues in each cube
-   LiAInd  index of the residue list 
+   LiAInd  index of the residue list
    CubAtX  cube number along x in which each receptor residue is
    CubAtY  cube number along y in which each receptor residue is
-   CubAtZ  cube number along z in which each receptor residue is 
+   CubAtZ  cube number along z in which each receptor residue is
    NuAtoC  number of residues in the current cube */
 {
   int i,j,k,l,LiAInd,*CubAtX,*CubAtY,*CubAtZ,NuAtoC;
@@ -28,10 +28,10 @@ void ReLAIC_en(int ReReNu,float **ReCoor,float GrSiCu_en,float *ReMinC,
   for (i=1;i<=ReReNu;i++) {
     CubAtX[i]=ffloor((ReCoor[AtReprRes[i]][1]-ReMinC[1])/GrSiCu_en)+1;
     CubAtY[i]=ffloor((ReCoor[AtReprRes[i]][2]-ReMinC[2])/GrSiCu_en)+1;
-    CubAtZ[i]=ffloor((ReCoor[AtReprRes[i]][3]-ReMinC[3])/GrSiCu_en)+1;    
+    CubAtZ[i]=ffloor((ReCoor[AtReprRes[i]][3]-ReMinC[3])/GrSiCu_en)+1;
   }
 
-/* Find the list of residues which are in each cube and the indexes for the 
+/* Find the list of residues which are in each cube and the indexes for the
    first and last residues */
   LiAInd=1;
 
@@ -44,7 +44,7 @@ void ReLAIC_en(int ReReNu,float **ReCoor,float GrSiCu_en,float *ReMinC,
 
         for (l=1;l<=ReReNu;l++) {
 
-          if ((CubAtX[l]==i)&&(CubAtY[l]==j)&&(CubAtZ[l]==k)) {            
+          if ((CubAtX[l]==i)&&(CubAtY[l]==j)&&(CubAtZ[l]==k)) {
             NuAtoC=NuAtoC+1;
             CubLiA_en[LiAInd]=l;
             LiAInd=LiAInd+1;
@@ -56,7 +56,7 @@ void ReLAIC_en(int ReReNu,float **ReCoor,float GrSiCu_en,float *ReMinC,
           CubFAI_en[i][j][k]=0;
           CubLAI_en[i][j][k]=0;
         }
-        else 
+        else
           CubLAI_en[i][j][k]=LiAInd-1;
 
       }
